added nfvitype to nsd nf template by making new j2 + changing how nfd processor works
